+++ module-api/src/main/java/org/example/controller/ReservationController.java
@@ -4,10 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.example.baseresponse.BaseResponse;
 import org.example.baseresponse.BaseResponseStatus;
 import org.example.dto.request.ReservationRequestDto;
-import org.example.service.ReservationService;
-import org.springframework.http.ResponseEntity;
+import org.example.service.reservation.ReservationService;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
